feat: delete comments cascade

+++ src/server/router/comment.router.ts
@@ -1,3 +1,4 @@
+import { Comment } from "@prisma/client";
 import { createRouter } from "@server/createRouter";
 import * as trpc from "@trpc/server";
 import {
@@ -100,18 +101,38 @@ export const commentRouter = createRouter()
       try {
         const { commentId } = input;
 
-        // TO-DO: Finish delete comment implementation
-        await ctx.prisma.comment.deleteMany({
-          where: {
-            parentId: commentId,
-          },
-        });
+        const deleteChildComments = async (commentId: string) => {
+          const oneLevelDownReplies = await ctx.prisma.comment.findMany({
+            where: {
+              parentId: commentId,
+            },
+          });
 
-        await ctx.prisma.comment.delete({
-          where: {
-            id: commentId,
-          },
-        });
+          // If no replies, delete comment.
+          if (!oneLevelDownReplies.length) {
+            return await ctx.prisma.comment.delete({
+              where: {
+                id: commentId,
+              },
+            });
+          }
+
+          // If has replies, check for other replies inside the replies.
+          if (oneLevelDownReplies.length > 0) {
+            for (const reply of oneLevelDownReplies) {
+              await deleteChildComments(reply.id);
+            }
+
+            // After checking all replies, delete comment.
+            await ctx.prisma.comment.delete({
+              where: {
+                id: commentId,
+              },
+            });
+          }
+        };
+
+        await deleteChildComments(commentId);
 
         return true;
       } catch (e) {
